Fix Growth Chamber deleting when using EG trees
Since the items were never cloned, the growth chamber was always referring to the original item. This is fine (although frowned upon) as long as you leave the ItemStack unchanged as it is Immutable, but with EG trees the amount is set to 3 which errors as you can't change an Immutable ItemStack.
Here is an error example of the above happening: https://pastebin.com/BkncFq96
Cloning the ItemStack will circumvent the issue, I cloned all of them for coherence.